requirements-assistant.txt
groq>=0.31.0           # 用雲端 Groq


# RAG 本地嵌入
sentence-transformers>=5.1.0
numpy>=2.3.2
python-dotenv>=1.1.1

djangorestframework>=3.16.0
django-cors-headers>=4.7.0